python/requirements.txt
flask==3.1.3
werkzeug==3.1.3
render_template==0.2.0
mysql-connector-python==9.6.0